partial rename of interfaces
