+++ packages/kbn-optimizer/limits.yml
@@ -112,7 +112,7 @@ pageLoadAssetSize:
   embeddableEnhanced: 22107
   uiActionsEnhanced: 38494
   urlDrilldown: 30063
-  indexPatternEditor: 19123
+  dataViewEditor: 19123
   indexPatternFieldEditor: 34448
   indexPatternManagement: 19165
   reporting: 57003
