--- src/plugins/data_view_editor/public/components/index_pattern_editor.scss
@@ -1,13 +1,0 @@
-.indexPatternEditor__form {
-  flex-grow: 1;
-}
-
-.fieldEditor__mainFlyoutPanel {
-  display: flex;
-  flex-direction: column;
-}
-
-.indexPatternEditor__footer {
-  margin-left: -$euiSizeL;
-  margin-right: -$euiSizeL;
-}
