+++ src/plugins/data_view_editor/public/types.ts
@@ -17,7 +17,7 @@ import {
 
 import { EuiComboBoxOptionOption } from '@elastic/eui';
 
-import type { IndexPattern } from 'src/plugins/data/public';
+import type { DataView } from 'src/plugins/data_views/public';
 import { DataPublicPluginStart, IndexPatternAggRestrictions } from './shared_imports';
 
 export interface IndexPatternEditorContext {
@@ -31,12 +31,12 @@ export interface IndexPatternEditorContext {
 }
 
 /** @public */
-export interface IndexPatternEditorProps {
+export interface DataViewEditorProps {
   /**
    * Handler for the "save" footer button
    * @param indexPattern - newly created index pattern
    */
-  onSave: (indexPattern: IndexPattern) => void;
+  onSave: (dataView: DataView) => void;
   /**
    * Handler for the "cancel" footer button
    */
@@ -55,10 +55,10 @@ export interface IndexPatternEditorProps {
 export interface PluginSetup {}
 
 export interface PluginStart {
-  openEditor(options: IndexPatternEditorProps): () => void;
-  IndexPatternEditorComponent: FC<IndexPatternEditorProps>;
+  openEditor(options: DataViewEditorProps): () => void;
+  IndexPatternEditorComponent: FC<DataViewEditorProps>;
   userPermissions: {
-    editIndexPattern: () => boolean;
+    editDataView: () => boolean;
   };
 }
